review round 6: soft-limit warnings + staleness comment + vendor sentinel + CI fixes
- database: warn! when fast_resolve_cache / delta_provider_cache cross
  the documented design target (10× soft limit = 10k entries) so the
  no-eviction caveat surfaces without needing scheduled timefusion_stats
  queries. Threshold-multiple cadence so log volume tracks tenant growth.
- database: comment the bounded-staleness window on the OnceCell-based
  delta_provider_cache — a task that captured the v=N cell before a
  flush bumped it sees one-query pre-flush data, then converges. Intentional
  for an append-only OLAP system; the alternative (re-check version after
  await) would defeat the cache's whole purpose.
- vendor/datafusion-postgres/PATCHES.md: add a CI grep sentinel that
  fails the build if a future cargo vendor / manual resync silently
  reverts our was_pre_optimized patches.

CI / clippy:
- membuffer_concurrency_bench: drop redundant 'as f64' on writer_rate
  (unnecessary_cast).
- database: use size.is_multiple_of(threshold) instead of % == 0
  (manual_is_multiple_of).
- database: drop redundant mut on the DashMap entry binding.

bollard / bollard-stubs in Cargo.lock are dev-only (testcontainers
under [dev-dependencies] — verified via cargo tree --edges normal,
'nothing to print'). Production binary is unaffected.

+++ tests/membuffer_concurrency_bench.rs
@@ -79,7 +79,7 @@ fn concurrent_insert_query_bench() {
         let inserts = inserts.clone();
         let pid = format!("proj-{p:04}");
         let per_batch_sleep = if writer_rate > 0.0 {
-            Duration::from_secs_f64(batch_size as f64 / writer_rate as f64)
+            Duration::from_secs_f64(batch_size as f64 / writer_rate)
         } else {
             Duration::ZERO
         };
